Moving make_store files in to store directory.
