+++ nano/core_test/block_store.cpp
@@ -5,12 +5,12 @@
 #include <nano/lib/utility.hpp>
 #include <nano/lib/work.hpp>
 #include <nano/node/common.hpp>
-#include <nano/node/make_store.hpp>
 #include <nano/secure/ledger.hpp>
 #include <nano/secure/utility.hpp>
 #include <nano/store/account.hpp>
 #include <nano/store/block.hpp>
 #include <nano/store/lmdb/lmdb.hpp>
+#include <nano/store/make_store.hpp>
 #include <nano/store/rocksdb/rocksdb.hpp>
 #include <nano/store/versioning.hpp>
 #include <nano/test_common/system.hpp>
