+++ nano/node/node.cpp
@@ -4,7 +4,6 @@
 #include <nano/lib/utility.hpp>
 #include <nano/node/common.hpp>
 #include <nano/node/daemonconfig.hpp>
-#include <nano/node/make_store.hpp>
 #include <nano/node/node.hpp>
 #include <nano/node/scheduler/component.hpp>
 #include <nano/node/scheduler/hinted.hpp>
@@ -14,6 +13,7 @@
 #include <nano/node/telemetry.hpp>
 #include <nano/node/websocket.hpp>
 #include <nano/store/component.hpp>
+#include <nano/store/make_store.hpp>
 #include <nano/store/rocksdb/rocksdb.hpp>
 
 #include <boost/property_tree/json_parser.hpp>
